// src/components/checkout/_shared/types/color.types.ts
export type TextColor =
  | 'strong'
  | 'weak'
  | 'brand'
  | 'brand-secondary'
  | 'disabled'
  | 'error'
  | 'warning'
  | 'success'
  | 'information'
  | 'inverse-strong'
  | 'inverse-weak'
  | 'inverse-disabled';

export type IconColor =
  | 'strong'
  | 'neutral'
  | 'weak'
  | 'brand'
  | 'brand-secondary'
  | 'disabled'
  | 'error'
  | 'warning'
  | 'success'
  | 'information'
  | 'inverse'
  | 'inverse-strong'
  | 'inverse-disabled';